Move execute Build to create_image.dart
