+++ lib/src/docker/create_image.dart
@@ -1,3 +1,4 @@
+import 'dart:io' as io;
 import 'package:dockerize_sidekick_plugin/src/util/command_runner.dart';
 import 'package:sidekick_core/sidekick_core.dart';
 
@@ -12,4 +13,27 @@ void createDockerImage(String environmentName) {
     workingDirectory: repository.root.directory('server'),
     successMessage: '✅ Created image ${mainProject!.name}:$environmentName 🎉',
   );
+}
+
+// Helper method which is under the hood calling the build command
+void executeBuild({
+  required bool buildContainer,
+  required String envName,
+  String? path,
+}) {
+  final entryPointPath = path ?? Repository.requiredEntryPoint.path;
+  final process = io.Process.runSync(
+    entryPointPath,
+    [
+      'docker',
+      'build',
+      '--env=$envName',
+      if (buildContainer) '--docker-only',
+    ],
+    runInShell: true,
+  );
+  print(process.stdout);
+  process.exitCode == 0
+      ? print(green('Build successful'))
+      : print(red('Build failed'));
 }
